fixup! Add wxUILocale::CmpLocalizedStrings function
