+++ include/wx/localedefs.h
@@ -112,17 +112,52 @@ struct WXDLLIMPEXP_BASE wxLanguageInfo
 // wxLocaleIdent: allows to fully identify a locale under all platforms
 // ----------------------------------------------------------------------------
 
-class wxLocaleIdent
+class WXDLLIMPEXP_BASE wxLocaleIdent
 {
 public:
-    wxLocaleIdent() { }                                  // Leave language empty
-    wxLocaleIdent(const wxString& language);             // Construct name from language
-    wxLocaleIdent& Language(const wxString& language);   // Set language
-    wxLocaleIdent& Region(const wxString& region);       // Set region
-    wxLocaleIdent& Script(const wxString& script);       // Set script
-    wxLocaleIdent& Modifier(const wxString& modifier);   // Set modifier
-    wxString GetName() const;                            // Construct platform dependent name
-    bool IsDefault() const;                              // Empty language represents user's default language
+    // Leave language empty
+    wxLocaleIdent() { }
+
+    // Construct name from language
+    wxLocaleIdent(const wxString& language) : m_language(language) { }
+
+    // Set language
+    wxLocaleIdent& Language(const wxString& language)
+    {
+        m_language = language;
+        return *this;
+    }
+
+    // Set region
+    wxLocaleIdent& Region(const wxString& region)
+    {
+        m_region = region;
+        return *this;
+    }
+
+    // Set script
+    wxLocaleIdent& Script(const wxString& script)
+    {
+        m_script = script;
+        return *this;
+    }
+
+    // Set modifier
+    wxLocaleIdent& Modifier(const wxString& modifier)
+    {
+        m_modifier = modifier;
+        return *this;
+    }
+
+    // Construct platform dependent name
+    wxString GetName() const;
+
+    // Empty language represents user's default language
+    bool IsDefault() const
+    {
+        return m_language.empty();
+    }
+
 private:
     wxString m_language;
     wxString m_region;
